Fix warnings (implicit conversion loses and use of deprecated function)

+++ core/2d/CCFontFNT.cpp
@@ -151,7 +151,7 @@ std::set<unsigned int>* BMFontConfiguration::parseConfigFile(std::string_view co
     if (data.size() >= (sizeof("BMP") - 1) && memcmp("BMF", data.c_str(), sizeof("BMP") - 1) == 0)
     {
         // Handle fnt file of binary format
-        std::set<unsigned int>* ret = parseBinaryConfigFile((unsigned char*)&data.front(), data.size(), controlFile);
+        std::set<unsigned int>* ret = parseBinaryConfigFile((unsigned char*)&data.front(), static_cast<uint32_t>(data.size()), controlFile);
         return ret;
     }
     if (data[0] == 0)
